# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for GameHub PS Rental Management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UnitStatus = (typeof UnitStatus)[keyof typeof UnitStatus];

export const UnitStatus = {
  available: "available",
  occupied: "occupied",
} as const;

export interface Unit {
  id: number;
  name: string;
  status: UnitStatus;
  /** Rate in IDR per hour */
  hourlyRate: number;
  createdAt: string;
}

export interface CreateUnitBody {
  name: string;
  hourlyRate: number;
}

export type UpdateUnitBodyStatus =
  (typeof UpdateUnitBodyStatus)[keyof typeof UpdateUnitBodyStatus];

export const UpdateUnitBodyStatus = {
  available: "available",
  occupied: "occupied",
} as const;

export interface UpdateUnitBody {
  name?: string;
  hourlyRate?: number;
  status?: UpdateUnitBodyStatus;
}

export interface Product {
  id: number;
  name: string;
  /** Price in IDR */
  price: number;
  stock: number;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  price: number;
  stock: number;
}

export interface UpdateProductBody {
  name?: string;
  price?: number;
  stock?: number;
}

export type RentalStatus = (typeof RentalStatus)[keyof typeof RentalStatus];

export const RentalStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface Rental {
  id: number;
  unitId: number;
  unitName: string;
  customerName: string;
  startTime: string;
  endTime?: string | null;
  durationMinutes?: number | null;
  /** Total cost in IDR */
  totalCost?: number | null;
  status: RentalStatus;
  createdAt: string;
}

export interface ActiveRental {
  id: number;
  unitId: number;
  unitName: string;
  customerName: string;
  startTime: string;
  elapsedMinutes: number;
  /** Estimated cost so far in IDR */
  estimatedCost: number;
  hourlyRate: number;
}

export interface StartRentalBody {
  unitId: number;
  customerName: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  rental: "rental",
  product: "product",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  description: string;
  /** Amount in IDR */
  amount: number;
  rentalId?: number | null;
  createdAt: string;
}

export interface CreateTransactionBody {
  productId: number;
  quantity?: number;
}

export interface DailyIncome {
  date: string;
  income: number;
}

export interface DashboardStats {
  /** Total income today in IDR */
  todayIncome: number;
  activeRentals: number;
  availableUnits: number;
  totalUnits: number;
  totalProducts: number;
  todayTransactions: number;
  weeklyIncome: DailyIncome[];
}

export interface DeleteResult {
  success: boolean;
  message: string;
}

export type ListRentalsParams = {
  status?: ListRentalsStatus;
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
};

export type ListRentalsStatus =
  (typeof ListRentalsStatus)[keyof typeof ListRentalsStatus];

export const ListRentalsStatus = {
  active: "active",
  completed: "completed",
} as const;

export type ListTransactionsParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
  type?: ListTransactionsType;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  rental: "rental",
  product: "product",
} as const;

export type ListRecentTransactionsParams = {
  limit?: number;
};