# lib/api-zod/src/generated/types/createUnitBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for GameHub PS Rental Management
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUnitBody {
  name: string;
  hourlyRate: number;
}