# lib/api-zod/src/generated/types/dailyIncome.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for GameHub PS Rental Management
 * OpenAPI spec version: 0.1.0
 */

export interface DailyIncome {
  date: string;
  income: number;
}