# lib/api-zod/src/generated/types/rental.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for GameHub PS Rental Management
 * OpenAPI spec version: 0.1.0
 */
import type { RentalStatus } from "./rentalStatus";

export interface Rental {
  id: number;
  unitId: number;
  unitName: string;
  customerName: string;
  startTime: Date;
  endTime?: Date | null;
  durationMinutes?: number | null;
  /** Total cost in IDR */
  totalCost?: number | null;
  status: RentalStatus;
  createdAt: Date;
}