# lib/api-zod/src/generated/types/unit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for GameHub PS Rental Management
 * OpenAPI spec version: 0.1.0
 */
import type { UnitStatus } from "./unitStatus";

export interface Unit {
  id: number;
  name: string;
  status: UnitStatus;
  /** Rate in IDR per hour */
  hourlyRate: number;
  createdAt: Date;
}